moved OCR to plugin
tests also updated to pass using Moto
(still awaiting release permission for fixtures)
